Move Color RGB string decoding to WooFoundation
